Several fix bugs and the Parallex Background
Parallex Background works with a moving object at runtime
The healt bar fix, the numbers were inverted
the dagger now pass through dead enemies
the reaper can again do its special attack
the ranger can be damaged always
