+++ Project Smash/Assets/SpineAnimationRun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d5ba82612c0c3143bcba7c79d5d6f18
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
