+++ Project Smash/Assets/_Game/Characters/Character.prefab
@@ -24,7 +24,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3784079572925901943}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 3.6575615, y: 5.886004, z: -11.016563}
+  m_LocalPosition: {x: 3.6575615, y: 5.886004, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 4664805819710454425}
