+++ Project Smash/Assets/_Game/Characters/Enemies/Reaper/ForGameUse/Reaper V2.2.3.1 (Red).prefab
@@ -14,6 +14,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
         type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
+        type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
@@ -29,6 +34,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
         type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
+        type: 3}
       propertyPath: m_LocalRotation.x
       value: 0
       objectReference: {fileID: 0}
@@ -40,16 +50,6 @@ PrefabInstance:
     - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
         type: 3}
       propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
-        type: 3}
-      propertyPath: m_RootOrder
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 7303274224679174719, guid: 0c2a3884945251e40894e55c55b503f7,
@@ -128,6 +128,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1999293970786120266, guid: 2568f9267010504479b246f256e48050,
         type: 3}
+      propertyPath: m_RootOrder
+      value: 3
+      objectReference: {fileID: 0}
+    - target: {fileID: 1999293970786120266, guid: 2568f9267010504479b246f256e48050,
+        type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
@@ -140,6 +145,11 @@ PrefabInstance:
         type: 3}
       propertyPath: m_LocalPosition.z
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1999293970786120266, guid: 2568f9267010504479b246f256e48050,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 1999293970786120266, guid: 2568f9267010504479b246f256e48050,
         type: 3}
@@ -158,16 +168,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1999293970786120266, guid: 2568f9267010504479b246f256e48050,
         type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1999293970786120266, guid: 2568f9267010504479b246f256e48050,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 3
-      objectReference: {fileID: 0}
-    - target: {fileID: 1999293970786120266, guid: 2568f9267010504479b246f256e48050,
-        type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
       value: 0
       objectReference: {fileID: 0}
@@ -183,23 +183,13 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
-      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.size
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
-        type: 3}
       propertyPath: onStunStateEnded.m_PersistentCalls.m_Calls.Array.size
       value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
-      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_Mode
+      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.size
       value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
-        type: 3}
-      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_CallState
-      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
@@ -208,8 +198,8 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
-      propertyPath: onStunStateEnded.m_PersistentCalls.m_Calls.Array.data[0].m_CallState
-      value: 2
+      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_Mode
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
@@ -223,13 +213,13 @@ PrefabInstance:
       objectReference: {fileID: 8956312689921795048}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
-      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
-      value: ShowPrompt
+      propertyPath: onStunStateEnded.m_PersistentCalls.m_Calls.Array.data[0].m_CallState
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
-      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_Arguments.m_ObjectArgumentAssemblyTypeName
-      value: UnityEngine.Object, UnityEngine
+      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_CallState
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
@@ -238,7 +228,17 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
         type: 3}
+      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
+      value: ShowPrompt
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
+        type: 3}
       propertyPath: onStunStateEnded.m_PersistentCalls.m_Calls.Array.data[0].m_Arguments.m_ObjectArgumentAssemblyTypeName
+      value: UnityEngine.Object, UnityEngine
+      objectReference: {fileID: 0}
+    - target: {fileID: 4136347048131046087, guid: 2568f9267010504479b246f256e48050,
+        type: 3}
+      propertyPath: onStunStateStart.m_PersistentCalls.m_Calls.Array.data[0].m_Arguments.m_ObjectArgumentAssemblyTypeName
       value: UnityEngine.Object, UnityEngine
       objectReference: {fileID: 0}
     - target: {fileID: 4748840381959370614, guid: 2568f9267010504479b246f256e48050,
