+++ Project Smash/Assets/_Game/Characters/Enemies/Reaper/NolForGameUse/Reaper Character V2.2.3.1 (Red).prefab
@@ -17,6 +17,11 @@ PrefabInstance:
       propertyPath: m_Mesh
       value: 
       objectReference: {fileID: 0}
+    - target: {fileID: 3048860714097551001, guid: 96e7ddd3c92e88f48907d75d38b1268e,
+        type: 3}
+      propertyPath: maxSlopeAngle
+      value: 30
+      objectReference: {fileID: 0}
     - target: {fileID: 3859284013105656982, guid: 96e7ddd3c92e88f48907d75d38b1268e,
         type: 3}
       propertyPath: fsm.variables.boolVariables.Array.data[1].value
@@ -26,6 +31,11 @@ PrefabInstance:
         type: 3}
       propertyPath: fsm.states.Array.data[4].actionData.fsmVarParams.Array.data[0].boolValue
       value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3916743646091414752, guid: 96e7ddd3c92e88f48907d75d38b1268e,
+        type: 3}
+      propertyPath: fsm.variables.boolVariables.Array.data[1].value
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 3916743646091414752, guid: 96e7ddd3c92e88f48907d75d38b1268e,
         type: 3}
@@ -57,15 +67,15 @@ PrefabInstance:
       propertyPath: fsm.variables.objectVariables.Array.data[8].typeName
       value: PSmash.Combat.EnemyFighter
       objectReference: {fileID: 0}
-    - target: {fileID: 3916743646091414752, guid: 96e7ddd3c92e88f48907d75d38b1268e,
-        type: 3}
-      propertyPath: fsm.variables.boolVariables.Array.data[1].value
-      value: 0
-      objectReference: {fileID: 0}
     - target: {fileID: 5429202279606030884, guid: 96e7ddd3c92e88f48907d75d38b1268e,
         type: 3}
       propertyPath: m_Name
       value: Reaper Character V2.2.3.1 (Red)
+      objectReference: {fileID: 0}
+    - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
         type: 3}
@@ -84,6 +94,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
         type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
+        type: 3}
       propertyPath: m_LocalRotation.x
       value: 0
       objectReference: {fileID: 0}
@@ -95,16 +110,6 @@ PrefabInstance:
     - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
         type: 3}
       propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
-        type: 3}
-      propertyPath: m_RootOrder
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 5429202279606030886, guid: 96e7ddd3c92e88f48907d75d38b1268e,
