+++ Project Smash/Assets/_Game/Characters/PlayerHolder.prefab
@@ -24,7 +24,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6228382161423451194}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0.36432192, y: 0.4884705, z: -0.55853075}
+  m_LocalPosition: {x: 0.36432192, y: 0.4884705, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 93117808021631179}
@@ -56,11 +56,6 @@ PrefabInstance:
     - target: {fileID: 6581303975160096540, guid: 3b552e7319322ac42a1da3aec1ed7c4b,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6581303975160096540, guid: 3b552e7319322ac42a1da3aec1ed7c4b,
-        type: 3}
-      propertyPath: m_LocalPosition.z
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 6581303975160096540, guid: 3b552e7319322ac42a1da3aec1ed7c4b,
