+++ Project Smash/Assets/_Game/Objects/Weapons/Dagger.prefab
@@ -131,7 +131,7 @@ GameObject:
   - component: {fileID: 5957079273200037515}
   - component: {fileID: 5957079273200037513}
   - component: {fileID: 3048807680942522201}
-  m_Layer: 0
+  m_Layer: 16
   m_Name: Dagger
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -285,7 +285,7 @@ AudioSource:
     m_Curve:
     - serializedVersion: 3
       time: 0
-      value: 1
+      value: 0
       inSlope: 0
       outSlope: 0
       tangentMode: 0
