+++ Project Smash/Assets/_Game/UI/ForGameUse/UI V0.2.prefab
@@ -137,9 +137,24 @@ PrefabInstance:
       propertyPath: m_AnchoredPosition.y
       value: -134.2002
       objectReference: {fileID: 0}
+    - target: {fileID: 498085438624982495, guid: 1b87b007cde5dc043a5acc4fea8f0c74,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 885513678742120174, guid: 1b87b007cde5dc043a5acc4fea8f0c74,
         type: 3}
       propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1583322601568676357, guid: 1b87b007cde5dc043a5acc4fea8f0c74,
+        type: 3}
+      propertyPath: m_textAlignment
+      value: 257
+      objectReference: {fileID: 0}
+    - target: {fileID: 1688596857664896324, guid: 1b87b007cde5dc043a5acc4fea8f0c74,
+        type: 3}
+      propertyPath: m_LocalPosition.y
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 2292325007850092986, guid: 1b87b007cde5dc043a5acc4fea8f0c74,
@@ -163,6 +178,11 @@ PrefabInstance:
       value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 3146155987117696717, guid: 1b87b007cde5dc043a5acc4fea8f0c74,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4281785514733263224, guid: 1b87b007cde5dc043a5acc4fea8f0c74,
         type: 3}
       propertyPath: m_LocalPosition.y
       value: 0
