+++ Project Smash/Assets/_Scenes/LVL1_Paramo/Zone_1.unity
@@ -1945,6 +1945,11 @@ PrefabInstance:
       propertyPath: m_Name
       value: Ranger01_01 (6)
       objectReference: {fileID: 0}
+    - target: {fileID: 6967242623932699408, guid: 1d7a2d03514ea2944b20e6181545be4b,
+        type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 6967242623932699409, guid: 1d7a2d03514ea2944b20e6181545be4b,
         type: 3}
       propertyPath: m_RootOrder
@@ -2344,12 +2349,12 @@ PrefabInstance:
     - target: {fileID: 1961288071010327238, guid: 0c292b7de4920264087b71a5ce9ec610,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: 865.21
+      value: 835
       objectReference: {fileID: 0}
     - target: {fileID: 1961288071010327238, guid: 0c292b7de4920264087b71a5ce9ec610,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: -42.21
+      value: -40.3
       objectReference: {fileID: 0}
     - target: {fileID: 1961288071010327238, guid: 0c292b7de4920264087b71a5ce9ec610,
         type: 3}
@@ -3064,6 +3069,16 @@ PrefabInstance:
       propertyPath: m_Mesh
       value: 
       objectReference: {fileID: 0}
+    - target: {fileID: 6666036672284579343, guid: 11d9e5e1d156cf741ae0f33164010757,
+        type: 3}
+      propertyPath: fsm.variables.objectVariables.Array.data[1].typeName
+      value: 
+      objectReference: {fileID: 0}
+    - target: {fileID: 6666036672284579343, guid: 11d9e5e1d156cf741ae0f33164010757,
+        type: 3}
+      propertyPath: fsm.variables.objectVariables.Array.data[6].typeName
+      value: 
+      objectReference: {fileID: 0}
     - target: {fileID: 7404332155987567964, guid: 11d9e5e1d156cf741ae0f33164010757,
         type: 3}
       propertyPath: m_RootOrder
@@ -3223,6 +3238,11 @@ PrefabInstance:
         type: 3}
       propertyPath: m_Name
       value: Ranger01_01 (9)
+      objectReference: {fileID: 0}
+    - target: {fileID: 6967242623932699408, guid: 1d7a2d03514ea2944b20e6181545be4b,
+        type: 3}
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 6967242623932699409, guid: 1d7a2d03514ea2944b20e6181545be4b,
         type: 3}
@@ -5809,6 +5829,26 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 638050037}
     m_Modifications:
+    - target: {fileID: 908321112579235340, guid: 1d7a2d03514ea2944b20e6181545be4b,
+        type: 3}
+      propertyPath: fsm.preprocessed
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 908321112579235340, guid: 1d7a2d03514ea2944b20e6181545be4b,
+        type: 3}
+      propertyPath: fsm.handleFixedUpdate
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3418587955931456289, guid: 1d7a2d03514ea2944b20e6181545be4b,
+        type: 3}
+      propertyPath: fsm.preprocessed
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3418587955931456289, guid: 1d7a2d03514ea2944b20e6181545be4b,
+        type: 3}
+      propertyPath: fsm.handleFixedUpdate
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 5596914603776198216, guid: 1d7a2d03514ea2944b20e6181545be4b,
         type: 3}
       propertyPath: m_Mesh
@@ -6845,95 +6885,6 @@ Transform:
   m_Father: {fileID: 771384506}
   m_RootOrder: 10
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &2036730095
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 11
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -6.2
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 3.6
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 240923723333461625, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_Name
-      value: Backgrounds Variant Level1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2091530393953771821, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_IsActive
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2344620000579340554, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_IsActive
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6084843065268732139, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_IsActive
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7516887770821948247, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
-        type: 3}
-      propertyPath: m_IsActive
-      value: 1
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 0293ac33fc1d6124e9c5ff1fb33729be, type: 3}
 --- !u!1001 &2042391892
 PrefabInstance:
   m_ObjectHideFlags: 0
