+++ Project Smash/Assets/_Scenes/Y_Temporal/Parallex Test 1.unity
@@ -98,7 +98,7 @@ LightmapSettings:
     m_TrainingDataDestination: TrainingData
     m_LightProbeSampleCountMultiplier: 4
   m_LightingDataAsset: {fileID: 0}
-  m_UseShadowmask: 1
+  m_UseShadowmask: 0
 --- !u!196 &4
 NavMeshSettings:
   serializedVersion: 2
@@ -1669,6 +1669,75 @@ Transform:
   m_Father: {fileID: 1762583203}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &896308537
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 14
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -6.2
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 3.6
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461624, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 240923723333461625, guid: 0293ac33fc1d6124e9c5ff1fb33729be,
+        type: 3}
+      propertyPath: m_Name
+      value: ParallexController Zone1
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 0293ac33fc1d6124e9c5ff1fb33729be, type: 3}
 --- !u!1 &919936891
 GameObject:
   m_ObjectHideFlags: 0
@@ -2108,95 +2177,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &5087308050262780652
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 5087308051029860104, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_Name
-      value: Backgrounds
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 14
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 3.4207208
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 11.936758
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -14.053247
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5087308051029860105, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7590738678276936735, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_Size.x
-      value: 29.222696
-      objectReference: {fileID: 0}
-    - target: {fileID: 7590738678276936735, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_Size.y
-      value: 23.976181
-      objectReference: {fileID: 0}
-    - target: {fileID: 7590738678276936735, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_Offset.x
-      value: -3.166007
-      objectReference: {fileID: 0}
-    - target: {fileID: 7590738678276936735, guid: e818c22070c0ae048836f6fa3cd5c85a,
-        type: 3}
-      propertyPath: m_Offset.y
-      value: 0.27136612
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: e818c22070c0ae048836f6fa3cd5c85a, type: 3}
 --- !u!1001 &8632366419996057637
 PrefabInstance:
   m_ObjectHideFlags: 0
