+++ Project Smash/Assets/_Scripts/SceneManagement/BackgroundLayers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90ada4a0091181945902e1f3f3d19d1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
