+++ Project Smash/Assets/test.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1928398600459f44c8208f2a5f00c598
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
