+++ Project Smash/ProjectSettings/EditorBuildSettings.asset
@@ -12,7 +12,7 @@ EditorBuildSettings:
     path: Assets/_Scenes/DingyMountains/DM_Entrance.unity
     guid: 54c35770e3adb5b42b99995a622d4e89
   - enabled: 0
-    path: Assets/_Scenes/Y_Temporal/Combat_General_Sandbox.unity
+    path: Assets/_Scenes/Y_Temporal/Enemy_GenericPurpose.unity
     guid: e20c1cd754452034981987a344c74b99
   - enabled: 0
     path: Assets/_Scenes/Y_Temporal/Combat_NonOptional_Sandbox.unity
