+++ Project Smash/ProjectSettings/Physics2DSettings.asset
@@ -53,4 +53,4 @@ Physics2DSettings:
   m_ColliderAsleepColor: {r: 0.5686275, g: 0.95686275, b: 0.54509807, a: 0.36078432}
   m_ColliderContactColor: {r: 1, g: 0, b: 1, a: 0.6862745}
   m_ColliderAABBColor: {r: 1, g: 1, b: 0, a: 0.2509804}
-  m_LayerCollisionMatrix: ff0f08ffff0f08ffff0f08ffffffffffff0f08ffff0f08ffffffffffffffffffff7f4affff0988ffff6929ffff271fffc8011fffc8ad49ffc805a9ffc82088ffc87c08ffc819a8ffc81808ffffff8fffc81800ffc84442ffc82120ffc8c28affffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff
+  m_LayerCollisionMatrix: ff0f08feff0f08feff0f08feffffffffff0f08feff0f08feffffffffffffffffff7f4affff0988ffff6929feff271ffec8011ffec8ad49fec805a9ffc82088ffc87c48fec819a8ffc81808feffff8fffc81800fec84442ffc82121fec8c28afec8c32afeffffffffffffffffffffffffffffffffffffffffffffffffffffffff
