+++ Project Smash/ProjectSettings/TagManager.asset
@@ -44,7 +44,7 @@ TagManager:
   - NotPlayerGround
   - EnemiesGhost
   - Cameras
-  - 
+  - Dead
   - 
   - 
   -
